requirements.txt
asgiref==3.2.10
attrs==20.1.0
Django==3.1
iniconfig==1.0.1
more-itertools==8.4.0
numpy==1.19.1
packaging==20.4
pandas==1.1.1
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==6.0.1
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
